[1607]: Move the UnvestedCoins GetLockedCoinsFn to its own file.
